refactor: 멤버 테스트 분리
